remove enums package. move some interfaces to root, others to internal
