+++ app/src/main/java/com/auth0/android/lock/app/DemoActivity.java
@@ -40,9 +40,9 @@ import com.auth0.android.lock.AuthenticationCallback;
 import com.auth0.android.lock.Lock;
 import com.auth0.android.lock.LockCallback;
 import com.auth0.android.lock.PasswordlessLock;
-import com.auth0.android.lock.enums.InitialScreen;
-import com.auth0.android.lock.enums.SocialButtonStyle;
-import com.auth0.android.lock.enums.UsernameStyle;
+import com.auth0.android.lock.InitialScreen;
+import com.auth0.android.lock.SocialButtonStyle;
+import com.auth0.android.lock.UsernameStyle;
 import com.auth0.android.lock.utils.LockException;
 import com.auth0.android.result.Credentials;
