+++ lib/src/main/java/com/auth0/android/lock/PasswordlessLockActivity.java
@@ -51,7 +51,7 @@ import android.widget.TextView;
 import com.auth0.android.authentication.AuthenticationAPIClient;
 import com.auth0.android.authentication.AuthenticationException;
 import com.auth0.android.lock.adapters.Country;
-import com.auth0.android.lock.enums.PasswordlessMode;
+import com.auth0.android.lock.internal.PasswordlessMode;
 import com.auth0.android.lock.errors.AuthenticationError;
 import com.auth0.android.lock.errors.LoginErrorMessageBuilder;
 import com.auth0.android.lock.events.CountryCodeChangeEvent;
