+++ lib/src/main/java/com/auth0/android/lock/internal/AuthMode.java
@@ -1,12 +1,12 @@
-package com.auth0.android.lock.enums;
+package com.auth0.android.lock.internal;
 
 import android.support.annotation.IntDef;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
-import static com.auth0.android.lock.enums.AuthMode.LOG_IN;
-import static com.auth0.android.lock.enums.AuthMode.SIGN_UP;
+import static com.auth0.android.lock.internal.AuthMode.LOG_IN;
+import static com.auth0.android.lock.internal.AuthMode.SIGN_UP;
 
 @IntDef({LOG_IN, SIGN_UP})
 @Retention(RetentionPolicy.SOURCE)
