+++ lib/src/main/java/com/auth0/android/lock/internal/AuthType.java
@@ -1,14 +1,14 @@
-package com.auth0.android.lock.enums;
+package com.auth0.android.lock.internal;
 
 import android.support.annotation.IntDef;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
-import static com.auth0.android.lock.enums.AuthType.DATABASE;
-import static com.auth0.android.lock.enums.AuthType.ENTERPRISE;
-import static com.auth0.android.lock.enums.AuthType.PASSWORDLESS;
-import static com.auth0.android.lock.enums.AuthType.SOCIAL;
+import static com.auth0.android.lock.internal.AuthType.DATABASE;
+import static com.auth0.android.lock.internal.AuthType.ENTERPRISE;
+import static com.auth0.android.lock.internal.AuthType.PASSWORDLESS;
+import static com.auth0.android.lock.internal.AuthType.SOCIAL;
 
 @IntDef({DATABASE, ENTERPRISE, PASSWORDLESS, SOCIAL})
 @Retention(RetentionPolicy.SOURCE)
