+++ lib/src/main/java/com/auth0/android/lock/views/ClassicLockView.java
@@ -42,7 +42,7 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 
 import com.auth0.android.lock.R;
-import com.auth0.android.lock.enums.InitialScreen;
+import com.auth0.android.lock.InitialScreen;
 import com.auth0.android.lock.events.DatabaseLoginEvent;
 import com.auth0.android.lock.events.DatabaseSignUpEvent;
 import com.auth0.android.lock.events.FetchApplicationEvent;
