+++ lib/src/main/java/com/auth0/android/lock/views/PasswordlessInputCodeFormView.java
@@ -33,15 +33,15 @@ import android.view.inputmethod.EditorInfo;
 import android.widget.TextView;
 
 import com.auth0.android.lock.R;
-import com.auth0.android.lock.enums.PasswordlessMode;
+import com.auth0.android.lock.internal.PasswordlessMode;
 import com.auth0.android.lock.events.PasswordlessLoginEvent;
 import com.auth0.android.lock.views.interfaces.LockWidgetPasswordless;
 
-import static com.auth0.android.lock.enums.PasswordlessMode.DISABLED;
-import static com.auth0.android.lock.enums.PasswordlessMode.EMAIL_CODE;
-import static com.auth0.android.lock.enums.PasswordlessMode.EMAIL_LINK;
-import static com.auth0.android.lock.enums.PasswordlessMode.SMS_CODE;
-import static com.auth0.android.lock.enums.PasswordlessMode.SMS_LINK;
+import static com.auth0.android.lock.internal.PasswordlessMode.DISABLED;
+import static com.auth0.android.lock.internal.PasswordlessMode.EMAIL_CODE;
+import static com.auth0.android.lock.internal.PasswordlessMode.EMAIL_LINK;
+import static com.auth0.android.lock.internal.PasswordlessMode.SMS_CODE;
+import static com.auth0.android.lock.internal.PasswordlessMode.SMS_LINK;
 
 public class PasswordlessInputCodeFormView extends FormView implements View.OnClickListener, TextView.OnEditorActionListener {
