+++ lib/src/main/java/com/auth0/android/lock/views/SocialView.java
@@ -32,7 +32,7 @@ import android.view.ViewGroup;
 import android.widget.LinearLayout;
 
 import com.auth0.android.lock.R;
-import com.auth0.android.lock.enums.AuthMode;
+import com.auth0.android.lock.internal.AuthMode;
 import com.auth0.android.lock.events.SocialConnectionEvent;
 import com.auth0.android.lock.internal.json.Connection;
 import com.auth0.android.lock.views.interfaces.LockWidgetSocial;
